backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
httpx==0.27.2
psycopg2-binary==2.9.7
SQLAlchemy==2.0.22
alembic==1.13.2
geoalchemy2==0.15.2